End of the Test Case. With notes :D
Few notes about my code

I created a custom Subsystem for session management using OnlineSubsystems session interface. However I did not created callback proxies (asyncronous blueprint nodes) because node callback proxies are created as classes in c++. It would be Ideal for this scenario where we should wait before taking action after a function call (As epic games did with the default Create Session, Join Session etc. nodes)

Also created a subsystem for managing User Interfaces. Event though it is not the best way to do it. It helps demonstrates Dependency Management and Allows me to handle everything from c++ in one place. Even though those are not quite efficient to do this way, I believe it demonstrates myunderstanding of c++ better. They would be problematic in a larger project because of the Subsystem I created is a GameInstanceSubsystem. I didn't dive in and created a perfect structure for this subsystem to live in GameInstance lifetime due to limited time.

ADoorBase and ALeverBase are the main actor of our game. They implement the main functionality and RPCs for interactions. I also Modified Third Person Character class for interactions over RPCs. Usually it's best to put gameplay rules into GameMode and keep track of them with GameStates. However I had to be fast due to my work and demonstrate as much as I can. I realized that I should have treat this mechanics as core gameplay mechanics. But I ran out of time.

I use Perforce for version control but since it's not possible to share with other people publicly available I uploaded the project to Git. I also share a shipping build with you via Google Drive.

About Project:

Main Menu:
You can click Host Game button to host a game On LAN and set the Session Name and Player Name(Set for every player) in this dialog. You can also set how many player you session allows. I capped it to 2 player minimum and 12 player maximum.
Find Session option works like in the valorant. It searches for available session in the backgroun and shows a Modal(Popup) when the query finishes. The modal closes after 5 seconds. If there is another session found in the last query it will automatically show you a new modal. Closed completely otherwise. Deny button on the Find Session works the same if there is another session it shows a new modal. Closed completely other wise. Puttin option to completely close the modals would be nice but I was running out of time so I passed it.
Quit Game shows you a verification popup and works as expected.

In Game:
One you join the game you can move around with WASD and Look around with mouse as expected. Intereactions with levers are done via "Left Mouse Click". Basically if an object is behind your character model and it has Interactable interface you can interact with it.
When you interact with the lever it broadcasts it state and Door response to it. I couldn't find or make a door rigged (That was the plan I was going to animate it) So right now door actor respond to you interaction via color change and a basic niagara particle. Visual aspects of those classes are handled in Blueprint bound to C++ delegates. Color of the door changes depending on its state. So, even though there is two different lever that our door is bounded (That is a option. In Unreal Project you can duplicate the lever then simply select the door and add a new array entry to Lever Array from details panel) it will be correctly replicated to all clients since door state is replicated as well as lever state.
By pressing the Escape button you can get back to main menu. Only problem there is when you get back to main menu session is not being destroyed (if your client is hosting). Upon creation of new Session I destroy if there is any session that carries the same name but should have been destroyed on leave. I couldn't finish it on time because of my day work.

About UMGs
Almost every UMG is derived from its respected c++ class and its functionalities are implemented in c++. Only the In game pause menu is pure blueprint.